# docs/swagger.yaml
basePath: /
definitions:
  models.APIKey:
    properties:
      createdAt:
        type: string
      id:
        type: string
      name:
        type: string
      secret:
        type: string
      updatedAt:
        type: string
      userId:
        type: string
    type: object
...
  models.CreateAPIKeyRequest:
    properties:
      name:
        type: string
    type: object
...
  models.CreateUserRequest:
    properties:
      avatar:
        description: Avatar is the user's avatar
        type: string
      email:
        description: |-
          Email is the user's email
          example: john.doe@example.com
        type: string
      name:
        description: |-
          Name is the user's name
          example: John Doe
        type: string
    type: object
  models.ErrorResponse:
    properties:
      code:
        type: string
      details: {}
      message:
        type: string
    type: object
...
  models.MessageResponse:
    properties:
      message:
        example: message
        type: string
    type: object
...
  models.UpdateUserRequest:
    properties:
      avatar:
        description: Avatar is the user's avatar
        type: string
      email:
        description: |-
          Email is the user's email
          example: john.doe@example.com
        type: string
      name:
        description: |-
          Name is the user's name
          example: John Doe
        type: string
    type: object
...
  models.UpgradeUserContractRequest:
    properties:
      newClassHash:
        description: |-
          NewClassHash is the class hash of the new implementation contract
          example: 0xabcdef1234567890abcdef1234567890abcdef1234567890abcdef1234567890
        type: string
    type: object
  models.UpgradeUserContractResponse:
    properties:
      transactionHash:
        description: |-
          TransactionHash is the hash of the upgrade transaction
          example: 0xabcdef1234567890abcdef1234567890abcdef1234567890abcdef1234567890
        type: string
    type: object
  models.User:
    properties:
      accountAddress:
        description: AccountAddress is the user's StarkNet account address
        type: string
      avatar:
        description: Avatar is the user's avatar
        type: string
      createdAt:
        description: CreatedAt is the time the user was created
        type: string
      email:
        description: |-
          Email is the user's email
          example: john.doe@example.com
        type: string
      id:
        description: |-
          ID is the user's ID
          example: 0x1234567890abcdef1234567890abcdef12345678
        type: string
      name:
        description: |-
          Name is the user's name
          example: John Doe
        type: string
      phoneNumber:
        description: |-
          PhoneNumber is the user's phone number
          example: 1234567890
        type: string
      privateKey:
        description: PrivateKey is the user's StarkNet private key
        type: string
      publicKey:
        description: PublicKey is the user's StarkNet public key
        type: string
      updatedAt:
        description: UpdatedAt is the time the user was last updated
        type: string
    type: object
host: localhost:8080
info:
  contact:
    email: support@infinirewards.io
    name: API Support
    url: http://www.infinirewards.io/support
  description: API for InfiniRewards - A Web3 Loyalty and Rewards Platform
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: InfiniRewards API
  version: "1.0"
paths:
...
  /user:
    delete:
      consumes:
      - application/json
      description: Delete authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: User deleted successfully
          schema:
            $ref: '#/definitions/models.MessageResponse'
        "401":
          description: Unauthorized access
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete user
      tags:
      - user
    get:
      consumes:
      - application/json
      description: Get authenticated user details
      produces:
      - application/json
      responses:
        "200":
          description: User details retrieved successfully
          schema:
            $ref: '#/definitions/models.User'
        "401":
          description: Authentication error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get user details
      tags:
      - user
    post:
      consumes:
      - application/json
      description: Create a new user
      parameters:
      - description: User Creation Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.CreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: User created successfully
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Invalid request format
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "409":
          description: User already exists
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      summary: Create user
      tags:
      - user
    put:
      consumes:
      - application/json
      description: Update authenticated user details
      parameters:
      - description: User Update Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.UpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: User updated successfully
          schema:
            $ref: '#/definitions/models.User'
        "400":
          description: Invalid request format
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized access
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update user
      tags:
      - user
  /user/api-keys:
    get:
      consumes:
      - application/json
      description: List all API keys for authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: List of API keys
          schema:
            items:
              $ref: '#/definitions/models.APIKey'
            type: array
        "401":
          description: Unauthorized access
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List API keys
      tags:
      - api-keys
    post:
      consumes:
      - application/json
      description: Create a new API key for authenticated user
      parameters:
      - description: API Key Creation Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.CreateAPIKeyRequest'
      produces:
      - application/json
      responses:
        "201":
          description: API key created successfully
          schema:
            $ref: '#/definitions/models.APIKey'
        "400":
          description: Invalid request format
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized access
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create API key
      tags:
      - api-keys
  /user/api-keys/{keyId}:
    delete:
      consumes:
      - application/json
      description: Delete an API key for authenticated user
      parameters:
      - description: API Key ID
        in: path
        name: keyId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: API key deleted successfully
          schema:
            $ref: '#/definitions/models.MessageResponse'
        "400":
          description: Invalid request format
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "401":
          description: Unauthorized access
          schema:
            $ref: '#/definitions/models.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete API key
      tags:
      - api-keys
  /user/upgrade:
    post:
      consumes:
      - application/json
      description: Upgrade a user contract
      parameters:
      - description: Upgrade User Contract Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/models.UpgradeUserContractRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.UpgradeUserContractResponse'
        "400":
          description: Bad Request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      security:
      - BearerAuth: []
      summary: Upgrade User Contract
      tags:
      - user
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Enter your bearer token in the format **Bearer <token>**
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"
...